requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astroid==2.4.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==1.2.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
isort==5.6.4
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain-core==0.1.1
langchain-google-genai==0.0.4
langsmith==0.0.71
lazy-object-proxy==1.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.5
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.15.1
pylint==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.0.2
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.1
rsa==4.9
six==1.15.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.2
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
wrapt==1.12.1
zipp==3.15.0
